Otat 522: Define assessment colors as enum

+++ src/main/java/org/flickit/flickitassessmentcore/application/port/in/assessment/CreateAssessmentCommand.java
@@ -19,12 +19,12 @@ public class CreateAssessmentCommand extends SelfValidating<CreateAssessmentComm
     @NotNull(message = CREATE_ASSESSMENT_ASSESSMENT_KIT_ID_NOT_NULL)
     Long assessmentKitId;
 
-    long colorId;
+    Integer colorId;
 
     @NotNull(message = CREATE_ASSESSMENT_SPACE_ID_NOT_NULL)
     Long spaceId;
 
-    public CreateAssessmentCommand(Long spaceId, String title, Long assessmentKitId, Long colorId) {
+    public CreateAssessmentCommand(Long spaceId, String title, Long assessmentKitId, Integer colorId) {
         this.title = title != null ? title.strip() : null;
         this.spaceId = spaceId;
         this.assessmentKitId = assessmentKitId;
